Fixed the issues with sorting\deleting
menu.h
Added a pragma once statement to 
Now the program uses macro statements instead of hardcoded parameters for the length of author, name and category in the Book structure.

menu.c
Fixed the unnecessary pointer to library size argument. We don't need to give the pointer as an argument because the value itself does not need to be (and cant be) changed. Now its just an int argument, not a int *.

Reworked the function for book deletion. Now it deletes books and shifts the array so all of non-empty books are in the beginning of the array. Now this function does not break the sorting function.

Rewrote the sorting function to get rid of a lot of similar code. Now instead of entire sorting functions implementations there is just one (qsort from the standard library) that uses different comparison functions, which reduces the amount of semi-idntical code and also makes the code shorter (because now the menu for sorting uses an array of comparison functions for qsort, which is shorter to write than an entire switch case)